Add Maven Framework; mattw.* -> io.mattw.*
